cs: tcs の f32 数値モデル (T226) へ追従 — double/long を粉砕
tcs 最新 (395c234) で double/long が TinyC# サブセット外になったため、
cs-lib と全 30 サンプルを int/float の 2 型へ移行する:

- double → float、実数リテラルへ f suffix 付与
- double を返す System.Math API は (float) cast で受ける (tcs 上流と同じ流儀)
- xorshift32 (lubx/Rand, 20_audio) は long + 0xFFFFFFFF マスクをやめ、
  int32 wrap + 算術シフト & マスクで Haxe と同じ乱数列をビット一致で維持
- manual 08-csharp を同期 (double/long 不可、struct/record struct 解禁)

tcs は float を Lua へそのまま実数として出す (f32 丸め・定数畳み込みなし)
ため生成 Lua の実行時挙動は不変。全 30 サンプルの WARP capture が移行前と
byte 一致することを確認済み (check/build/実行も 30/30)。

+++ cs-lib/lubx/Camera2d.cs
@@ -8,14 +8,14 @@
 /// </summary>
 public class Camera2d
 {
-    public double ppm;
-    public double originX;
-    public double originY;
-    public double logicalW;
-    public double logicalH;
+    public float ppm;
+    public float originX;
+    public float originY;
+    public float logicalW;
+    public float logicalH;
 
-    public Camera2d(double logicalW, double logicalH, double ppm,
-        double originX, double originY)
+    public Camera2d(float logicalW, float logicalH, float ppm,
+        float originX, float originY)
     {
         this.logicalW = logicalW;
         this.logicalH = logicalH;
@@ -25,25 +25,25 @@ public class Camera2d
     }
 
     /// <summary>world x → screen x</summary>
-    public double sx(double wx)
+    public float sx(float wx)
     {
         return originX + wx * ppm;
     }
 
     /// <summary>world y → screen y</summary>
-    public double sy(double wy)
+    public float sy(float wy)
     {
         return originY - wy * ppm;
     }
 
     /// <summary>screen x → world x</summary>
-    public double wx(double sxv)
+    public float wx(float sxv)
     {
         return (sxv - originX) / ppm;
     }
 
     /// <summary>screen y → world y</summary>
-    public double wy(double syv)
+    public float wy(float syv)
     {
         return (originY - syv) / ppm;
     }
